% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRate}
\alias{PyRate}
\title{PyRate}
\usage{
PyRate(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRate
}
\examples{
library(outsider)
PyRate <- module_import('PyRate', repo = 'dombennett/om..pyrate')
PyRate('-h')
}
