% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRateContinuous}
\alias{PyRateContinuous}
\title{PyRateContinuous}
\usage{
PyRateContinuous(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRateContinuous
}
\examples{
library(outsider)
PyRateContinuous <- module_import('PyRateContinuous',
                                  repo = 'dombennett/om..pyrate..2.0')
PyRateContinuous('-h')
}
