% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRateContinuousShift}
\alias{PyRateContinuousShift}
\title{PyRateContinuousShift}
\usage{
PyRateContinuousShift(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRateContinuousShift
}
\examples{
library(outsider)
PyRateContinuousShift <- module_import('PyRateContinuousShift',
                                       repo = 'dombennett/om..pyrate..2.0')
PyRateContinuousShift('-h')
}
