% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRateDES}
\alias{PyRateDES}
\title{PyRateDES}
\usage{
PyRateDES(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRateDES
}
\examples{
library(outsider)
PyRateDES <- module_import('PyRateDES', repo = 'dombennett/om..pyrate')
PyRateDES('-h')
}
