% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRateDES2}
\alias{PyRateDES2}
\title{PyRateDES2}
\usage{
PyRateDES2(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRateDES2
}
\examples{
library(outsider)
PyRateDES2 <- module_import('PyRateDES2', repo = 'dombennett/om..pyrate..2.0')
PyRateDES2('-h')
}
