% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRateMBD}
\alias{PyRateMBD}
\title{PyRateMBD}
\usage{
PyRateMBD(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRateMBD
}
\examples{
library(outsider)
PyRateMBD <- module_import('PyRateMBD', repo = 'dombennett/om..pyrate..2.0')
PyRateMBD('-h')
}
