% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRateMCDD}
\alias{PyRateMCDD}
\title{PyRateMCDD}
\usage{
PyRateMCDD(...)
}
\arguments{
\item{...}{Arguments}
}
\description{
Run PyRateMCDD
}
\examples{
library(outsider)
PyRateMCDD <- module_import('PyRateMCDD', repo = 'dombennett/om..pyrate')
PyRateMCDD('-h')
}
