% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PyRate}
\alias{PyRate}
\title{PyRate}
\usage{
PyRate(h = FALSE, v = FALSE, seed = NULL, useCPPlib = NULL,
  cite = FALSE, j = NULL, trait = NULL, logT = NULL, N = NULL,
  wd = NULL, out = NULL, singleton = NULL,
  frac_sampled_singleton = NULL, rescale = NULL, translate = NULL,
  d = NULL, clade = NULL, trait_file = NULL, restore_mcmc = NULL,
  filter = NULL, filter_taxa = NULL, initDiv = NULL,
  PPmodeltest = NULL, log_marginal_rates = NULL, tree = NULL,
  sampling = NULL, bdc = NULL, eqr = NULL, plot = NULL,
  plot2 = NULL, plot3 = NULL, plotRJ = NULL, grid_plot = NULL,
  root_plot = NULL, tag = NULL, ltt = NULL, mProb = NULL,
  BF = NULL, data_info = NULL, SE_stats = NULL, combLog = NULL,
  resample = NULL, col_tag = NULL, check_names = NULL,
  reduceLog = NULL, n = NULL, s = NULL, p = NULL, b = NULL,
  thread = NULL, A = NULL, use_DA = NULL, r = NULL, t = NULL,
  sw = NULL, M = NULL, B = NULL, T = NULL, S = NULL, k = NULL,
  a = NULL, dpp_f = NULL, dpp_hp = NULL, dpp_eK = NULL,
  dpp_grid = NULL, dpp_nB = NULL, rj_pr = NULL, rj_Ga = NULL,
  rj_Gb = NULL, rj_beta = NULL, rj_dm = NULL, rj_bd_shift = NULL,
  se_gibbs = NULL, pL = NULL, pM = NULL, pP = NULL, pS = NULL,
  pC = NULL, cauchy = NULL, mHPP = NULL, mL = NULL, mM = NULL,
  mC = NULL, mCov = NULL, mG = NULL, mPoiD = NULL, mDeath = NULL,
  mBDI = NULL, ncat = NULL, fixShift = NULL, qShift = NULL,
  fixSE = NULL, ADE = NULL, discrete = NULL, twotrait = NULL,
  bound = NULL, edgeShift = NULL, qFilter = NULL, tT = NULL,
  nT = NULL, tQ = NULL, tR = NULL, tS = NULL, fR = NULL,
  fS = NULL, fQ = NULL, tC = NULL, fU = NULL, multiR = NULL,
  tHP = NULL)
}
\arguments{
\item{h}{show this help message and exit}

\item{v}{show program's version number and exit}

\item{seed}{-1 random seed}

\item{useCPPlib}{1 Use C++ library if available (boolean)}

\item{cite}{print PyRate citation}

\item{j}{1 number of data set in input file}

\item{trait}{1 number of trait for Cov model}

\item{logT}{0 Transform trait (or rates for -plotRJ): 0) False, 1)
Ln(x), 2) Log10(x)}

\item{N}{N number of exant species}

\item{wd}{WD path to working directory}

\item{out}{OUT output tag}

\item{singleton}{0 Remove singletons (min no. occurrences)}

\item{frac_sampled_singleton}{0 Random fraction of singletons not removed}

\item{rescale}{1 Rescale data (e.g. -rescale 1000: 1 -> 1000,
time unit = 1Ky)}

\item{translate}{0 Shift data (e.g. -translate 10: 1My -> 10My)}

\item{d}{<input file> Load SE table}

\item{clade}{-1 clade analyzed (set to -1 to analyze all species)}

\item{trait_file}{<input file> Load trait table}

\item{restore_mcmc}{<input file> Load mcmc.log file}

\item{filter}{inf inf Filter lineages with all occurrences within time range}

\item{filter_taxa}{taxa_file Filter lineages within list (drop all others)}

\item{initDiv}{0 Number of initial lineages (option only available with
-d SE_table or -fixSE)}

\item{PPmodeltest}{Likelihood testing among preservation models}

\item{log_marginal_rates}{-1 0) save summary file, default for -A 4; 1)
save marginal rate file, default for -A 0,2}

\item{tree}{Tree file (NEXUS format)}

\item{sampling}{1.0 Taxon sampling (phylogeny)}

\item{bdc}{Run BDC:Compatible model}

\item{eqr}{Run BDC:Equal rate model}

\item{plot}{<input file> RTT plot (type 1): provide path to
'marginal_rates.log' files or 'marginal_rates' file}

\item{plot2}{<input file> RTT plot (type 2): provide path to
'marginal_rates.log' files or 'marginal_rates' file}

\item{plot3}{<input file> RTT plot for fixed number of shifts:
provide 'mcmc.log' file}

\item{plotRJ}{<input file> RTT plot for runs with '-log_marginal_rates 0':
provide path to 'mcmc.log' files}

\item{grid_plot}{1.0 Plot resolution in Myr (only for plot3 and plotRJ
commands)}

\item{root_plot}{0 User-defined root age for RTT plots}

\item{tag}{<*tag*.log> Tag identifying files to be combined and
plotted (-plot and -plot2) or summarized in SE table (-ginput)}

\item{ltt}{0 1) Plot lineages-through-time; 2) plot Log10(LTT)}

\item{mProb}{MPROB Input 'mcmc.log' file}

\item{BF}{<2 input files> [<2 input files> ...] Input
'marginal_likelihood.txt' files}

\item{data_info}{Summary information about an input data}

\item{SE_stats}{<extinction rate at the present, bin_size, #_simulations>
[<extinction rate at the present, bin_size, #_simulations> ...] Calculate
and plot stats from SE table: -ginput <path_to_mcmc.log> generate SE table
from *mcmc.log files}

\item{combLog}{<path_to_log_files> Combine (and resample) log files}

\item{resample}{0 Number of samples for each log file (-combLog). Use 0 to
keep all samples.}

\item{col_tag}{column names [column names ...] Columns to be combined using
combLog}

\item{check_names}{<*_SpeciesList.txt file> Automatic check for typos in
taxa names (provide SpeciesList file)}

\item{reduceLog}{<*_mcmc.log file> Reduce file size (mcmc.log) to quickly
assess convergence}

\item{n}{10000000 mcmc generations}

\item{s}{1000 sample freq.}

\item{p}{1000 print freq.}

\item{b}{0 burnin}

\item{thread}{4 4 no. threads used for BD and NHPP likelihood respectively
(set to 0 to bypass multi-threading)}

\item{A}{4 0) parameter estimation, 1) marginal likelihood, 2) BDMCMC, 3)
DPP, 4) RJMCMC}

\item{use_DA}{Use data augmentation for NHPP likelihood opf extant taxa}

\item{r}{1 MC3 - no. MCMC chains}

\item{t}{0.03 MC3 - temperature}

\item{sw}{100 MC3 - swap frequency}

\item{M}{10 BDMCMC/RJMCMC - frequency of model update}

\item{B}{1 BDMCMC - birth rate}

\item{T}{1.0 BDMCMC - time of model update}

\item{S}{1000 BDMCMC - start model update}

\item{k}{10 TI - no. scaling factors}

\item{a}{0.3 TI - shape beta distribution}

\item{dpp_f}{500 DPP - frequency}

\item{dpp_hp}{2.0 DPP - shape of gamma HP on concentration parameter}

\item{dpp_eK}{2.0 DPP - expected number of rate categories}

\item{dpp_grid}{1.5 DPP - size of time bins}

\item{dpp_nB}{0 DPP - number of time bins}

\item{rj_pr}{1 RJ - proposal (0: Gamma, 1: Weighted mean)}

\item{rj_Ga}{1.5 RJ - shape of gamma proposal (if rj_pr 0)}

\item{rj_Gb}{3.0 RJ - rate of gamma proposal (if rj_pr 0)}

\item{rj_beta}{10 RJ - shape of beta multiplier (if rj_pr 1)}

\item{rj_dm}{0 RJ - allow double moves (0: no, 1: yes)}

\item{rj_bd_shift}{0.5 RJ - 0: only sample shifts in speciation; 1:
only sample shifts in extinction}

\item{se_gibbs}{Use aprroximate S/E Gibbs sampler}

\item{pL}{1.1 1.1 Prior - speciation rate (Gamma <shape, rate>) |
(if shape=n,rate=0 -> rate estimated)}

\item{pM}{1.1 1.1 Prior - extinction rate (Gamma <shape, rate>) |
(if shape=n,rate=0 -> rate estimated)}

\item{pP}{1.5 1.5 Prior - preservation rate (Gamma <shape, rate>) |
(if shape=n,rate=0 -> rate estimated)}

\item{pS}{2.5 Prior - time frames (Dirichlet <shape>)}

\item{pC}{1 Prior - Covar parameters (Normal <standard deviation>) |
(if pC=0 -> sd estimated)}

\item{cauchy}{-1 -1 Prior - use hyper priors on sp/ex rates
(if 0 -> estimated)}

\item{mHPP}{Model - Homogeneous Poisson process of preservation}

\item{mL}{1 Model - no. (starting) time frames (speciation)}

\item{mM}{1 Model - no. (starting) time frames (extinction)}

\item{mC}{Model - constrain time frames (l,m)}

\item{mCov}{0 COVAR model: 1) speciation, 2) extinction, 3) speciation &
extinction, 4) preservation, 5) speciation & extinction & preservation}

\item{mG}{Model - Gamma heterogeneity of preservation rate}

\item{mPoiD}{Poisson-death diversification model}

\item{mDeath}{Pure-death model}

\item{mBDI}{-1 BDI sub-model - 0) birth-death, 1) immigration-death}

\item{ncat}{4 Model - Number of categories for Gamma heterogeneity}

\item{fixShift}{<input file> Input tab-delimited file}

\item{qShift}{<input file> Poisson process of preservation with shifts
(Input tab-delimited file)}

\item{fixSE}{<input file> Input mcmc.log file}

\item{ADE}{0 ADE model: 0) no age dependence 1) estimated age dep}

\item{discrete}{Discrete-trait-dependent BD model (requires -trait_file)}

\item{twotrait}{Discrete-trait-dependent extinction + Covar}

\item{bound}{0 0 Bounded BD model}

\item{edgeShift}{0 0 Fixed times of shifts at the edges (when -mL/-mM > 3)}

\item{qFilter}{1 if set to zero all shifts in preservation rates are kept,
even if outside observed timerange}

\item{tT}{1.0 Tuning - window size (ts, te)}

\item{nT}{5 Tuning - max number updated values (ts, te)}

\item{tQ}{TQ TQ Tuning - window sizes (q/alpha: 1.2 1.2)}

\item{tR}{1.2 Tuning - window size (rates)}

\item{tS}{1.0 Tuning - window size (time of shift)}

\item{fR}{0.5 Tuning - fraction of updated values (rates)}

\item{fS}{0.7 Tuning - fraction of updated values (shifts)}

\item{fQ}{0.5 Tuning - fraction of updated values (q rates, TPP)}

\item{tC}{TC TC TC Tuning - window sizes cov parameters (l,m,q)}

\item{fU}{FU FU FU Tuning - update freq. (q: .02, l/m: .18, cov: .08)}

\item{multiR}{1 Tuning - Proposals for l/m: 0) sliding win 1) muliplier}

\item{tHP}{THP THP Tuning - window sizes hyperpriors on l and m}
}
\description{
Run PyRate
}
